Move blog page files to views directory
